move answer toggle tests
