+++ src/main/webapp/app/shared/metis/posting-reactions-bar/post-reactions-bar/post-reactions-bar.component.html
@@ -67,14 +67,14 @@
     </ng-template>
     <!-- start discussion -->
     <div class="col-auto p-0" *ngIf="sortedAnswerPosts.length === 0">
-        <button class="reaction-button" (click)="openPostingCreateEditModal.emit()">
+        <button id="startDiscussionButton" class="reaction-button" (click)="openPostingCreateEditModal.emit()">
             <ngx-emoji class="emoji" [emoji]="speechBalloonId" [size]="16"></ngx-emoji>
             <span class="emoji-count">{{ 'artemisApp.metis.startDiscussion' | artemisTranslate }}</span>
         </button>
     </div>
     <!-- expand answers -->
     <div class="col-auto p-0" *ngIf="!showAnswers && sortedAnswerPosts.length > 0">
-        <button class="reaction-button" (click)="showAnswersChange.emit(true)">
+        <button id="expandAnswersButton" class="reaction-button" (click)="showAnswersChange.emit(true)">
             <ngx-emoji class="emoji" [emoji]="speechBalloonId" [size]="16"></ngx-emoji>
             <span class="emoji-count">{{
                 sortedAnswerPosts.length === 1
@@ -85,7 +85,7 @@
     </div>
     <!-- collapse answers -->
     <div class="col-auto p-0" *ngIf="showAnswers && sortedAnswerPosts.length > 0">
-        <button class="reaction-button clickable" [class.reaction-button--reacted]="showAnswers" (click)="showAnswersChange.emit(false)">
+        <button id="collapseAnswersButton" class="reaction-button clickable" [class.reaction-button--reacted]="showAnswers" (click)="showAnswersChange.emit(false)">
             <ngx-emoji class="emoji" [emoji]="closeCrossId" [size]="16"></ngx-emoji>
             <span class="emoji-count">{{ 'artemisApp.metis.collapseAnswers' | artemisTranslate }}</span>
         </button>
